Small refactoring and better type handling.
