+++ test-function/src/main.rs
@@ -1,108 +1,15 @@
+mod shm;
+
 use std::collections::HashSet;
 
+use crate::shm::rm_outliers_extern;
 use anna_api::lattice::{Lattice, SetLattice};
-use essa_api::deltalake_save;
-use essa_test_function::{
-    append_foo, repeat_string_extern, to_uppercase_extern,
-};
+use essa_test_function::{append_foo, repeat_string_extern, to_uppercase_extern};
 
 fn main() {
-    let to_get_acquisitions = essa_api::datafusion_run(
-        "SELECT frequencia, resistencia, temperatura FROM aquisicoes",
-        "/home/ceciliacsilva/Desktop/delta-rs/shm",
-    )
-    .unwrap();
-
-    println!("Testing R integration!");
-
-    let batch_size = essa_api::run_r(
-        "function(pzt_batch_measures){
-            return(length(pzt_batch_measures$resistencia))
-}",
-        &[to_get_acquisitions.0],
-    ).unwrap();
-
-    let temperature = essa_api::run_r(
-        "function(pzt_batch_measures){
-            return(pzt_batch_measures$temperatura)
-}",
-        &[to_get_acquisitions.0],
-    ).unwrap();
-
-    let frequency_points_qnt = essa_api::run_r(
-        "function(pzt_batch_measures){
-            return(length(pzt_batch_measures$frequencia[[1]]))
-}",
-        &[to_get_acquisitions.0],
-    ).unwrap();
-
-    let to_measures = essa_api::run_r(
-        r#"function(pzt_batch_measures, batch_size, frequency_points_qnt){
-  batch_size <- unlist(batch_size)[[1]]
-  frequency_points_qnt <- unlist(frequency_points_qnt)[[1]]
-  impedance_r <- pzt_batch_measures$resistencia
-  frequency_points <- pzt_batch_measures$frequencia[[1]]
-
-  measure_matrix <- matrix(NA, frequency_points_qnt, batch_size)
-
-  for(i in 1:batch_size){
-    measure_matrix[,i]<-impedance_r[[i]]
-  }
-  return(measure_matrix)
-}"#,
-        &[to_get_acquisitions.0, batch_size.0, frequency_points_qnt.0],
-    )
-    .unwrap();
-
-    let to_median_impedance = essa_api::run_r(
-        r#"function(measure_df, batch_size, frequency_points_qnt){
-  batch_size <- unlist(batch_size)[[1]]
-  frequency_points_qnt <- unlist(frequency_points_qnt)[[1]]
-  median_impedance_r = vector()
-  measure_matrix = matrix(unlist(measure_df), ncol=batch_size)
-
-  for(i in 1:frequency_points_qnt){
-    median_impedance_r[i]=median(as.numeric(measure_matrix[i,1:batch_size]))
-  }
-
-  return(median_impedance_r)
-}"#,
-        &[to_measures.0, batch_size.0, frequency_points_qnt.0],
-    )
-    .unwrap();
-
-    let _ = deltalake_save(
-        "/home/ceciliacsilva/Desktop/shm/median",
-        &[to_median_impedance.0],
-    );
-
-    let metric_calculation = essa_api::run_r(
-    r#"function(measure_matrix, median_impedance_r, batch_size, frequency_points_qnt){
-  batch_size <- unlist(batch_size)[[1]]
-  frequency_points_qnt <- unlist(frequency_points_qnt)[[1]]
-  metric_vector = vector()
-  CCD = vector()
-  measure_matrix = matrix(unlist(measure_matrix), ncol = batch_size)
-  median_impedance_r = matrix(unlist(measure_matrix), ncol=1)
-
-  for(i in 1:batch_size){
-    metric_vector[i] = sqrt(  sum(  (measure_matrix[,i] - median_impedance_r)^2/frequency_points_qnt  ))
-
-    CCD[i] = 1 - sum( (measure_matrix[,i] - mean(measure_matrix[,i])) *
-                          (median_impedance_r - mean(median_impedance_r)) /
-                                  (sd(measure_matrix[,i]) * sd(median_impedance_r)) ) / frequency_points_qnt
-  }
-
-  return(metric_vector)
-}"#,
-    &[to_measures.0, to_median_impedance.0, batch_size.0, frequency_points_qnt.0]
-    ).unwrap();
-
-    // TODO: this should be better.
-    let _ = deltalake_save(
-        "/home/ceciliacsilva/Desktop/shm/result",
-        &[metric_calculation.0, temperature.0],
-    );
+    println!("Removing outliers");
+    let removedp = rm_outliers_extern(13, 0).unwrap();
+    println!("Outliers removed?; {:?}", removedp.get().unwrap());
 
     println!("Hello world from test function!");
     let result = to_uppercase_extern("foobar".into()).expect("extern function call failed");
